feat: foto default para usuarios registrados

+++ hombresLoboLaravel/app/Http/Controllers/CloudinaryController.php
@@ -12,9 +12,10 @@ class CloudinaryController extends Controller
         $request->validated();
         //Al parecer laravel automaticamente revisa el token y te devuelve el usuario, cuando termine la funcion voy a probarlo
         $usuario = $request->user();
+        $foto_default = 'https://res.cloudinary.com/dj2m9tuoz/image/upload/v1763404018/default_user_photo_mxc7ra.jpg';
 
-        //Si el usuario tiene una foto se la borramos para dejar hueco en la bbdd
-        if ($usuario->foto_perfil) {
+        //Si el usuario tiene una foto se la borramos para dejar hueco en la bbdd, y ademas comprobamos que no sea la default para no borrarla
+        if ($usuario->foto_perfil && $usuario->foto_perfil != $foto_default) {
             Cloudinary::destroy($usuario->foto_perfil);
         }
